Create sendmessage
Function that end-to-end encrypts outgoing messages. If in a Quantum-Safe One Time Pad (QS-OTP) conversation, message is first QS-OTP end-to-end encrypted before that layer is wrapped inside of an AES-256 GCM end-to-end encrypted layer that uses Curve 25519 for key exchange. This layer is then wrapped in TLS v1.3 between app and server for a total of 3 independent layers of encryption, 2 of which are end-to-end.